Move state promela generation into promela_builder
* Move `State.generate_promela` to `PromelaBuilder._generate_state_promela`
* Create `@property` functions for accessing `State` fields
* Move associated tests to `test_promela_builder.py`

Closes #255
